requirements.txt
streamlit
goog-generativeai
python-dotenv
